alfred/requirements.txt
hass-client
litellm
openai
aiosqlite
python-dotenv
